Added get dataset info method in info controller (#1280)
